requirements.txt
pyTelegramBotAPI==4.12.0
pandas==2.0.3
python-dotenv==1.0.0 